+++ src/main/java/com/dongjin/tastingnote/alcohol/dto/AlcoholRequestCreateRequest.java
@@ -1,8 +1,6 @@
 package com.dongjin.tastingnote.alcohol.dto;
 
 import com.dongjin.tastingnote.alcohol.entity.AlcoholCategory;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Getter;
 
@@ -11,9 +9,10 @@ import java.util.List;
 @Getter
 public class AlcoholRequestCreateRequest {
 
-    @NotBlank
+    @Size(max = 100)
     private String name;
 
+    @Size(max = 100)
     private String nameKo;
 
     @Size(max = 10)
@@ -22,6 +21,5 @@ public class AlcoholRequestCreateRequest {
     @Size(max = 500)
     private String reason;
 
-    @NotNull
     private AlcoholCategory category;
 }