+++ src/main/java/com/dongjin/tastingnote/alcohol/entity/AlcoholRequestStatus.java
@@ -1,5 +1,5 @@
 package com.dongjin.tastingnote.alcohol.entity;
 
 public enum AlcoholRequestStatus {
-    PENDING, APPROVED, MERGED, REJECTED
+    PENDING, APPROVED, REJECTED
 }