+++ src/main/java/com/dongjin/tastingnote/alcohol/entity/AlcoholRequestType.java
@@ -1,0 +1,5 @@
+package com.dongjin.tastingnote.alcohol.entity;
+
+public enum AlcoholRequestType {
+    NEW, ALIAS
+}